Fixed error message on EOF in repl
Current behaviour is for dodo's repl to print an error message whenever it fails to read a line from stdin.
However, fgets reports failure by returning NULL, and it is up to us to decide whether it was simply EOF or a true error.

dodo did indeed check the EOF status of stdin, and set the exit code appropriately, but it printed
the error message regardless.

This commit alters this behaviour such that the error message is print only when a non-EOF error occurs.
The exit code behavious is left unaltered.