Fixed escape character bug, implemented distinction between implicit and explicit quit
The escape character required not one, but two characters after it, and it skipped over both of them.
Apart from just being wrong, this meant that `e/\//` failed, as it was being parsed as a string // which
wasn't correctly terminated.
Fixed this, behaviour and tested with `e/\//` and stuff like `e/\te\s\t/`, seems to work fine.
(GitHub Issue #18)

Added a separate return code for implicit and explicity quits in `execute()`.
This allowed the repl to break when an explicit quit was entered by the user, fixing GitHub Issue #16
Also needed to fix an old, undetected bug in parse() where the caching of `strlen(source)` became
misrepresentative if an escape character was parsed in parse_string. This bug was causing explicit quits to
be tacked onto the end of instructions containing an escape character in their string argument, and was
causing me headaches trying to implement the new repl quit behaviour.